+++ zmanim_bot/exceptions.py
@@ -51,3 +51,5 @@ class PolarCoordinatesException(ZmanimBotBaseException):
     ...
 
 
+class UnknownProcessorException(ZmanimBotBaseException):
+    ...
