--- zmanim_bot/processors/base_processor.py
@@ -1,5 +1,0 @@
-from abc import ABC
-
-
-class BaseProcessor(ABC):
-    ...
